Move follow title/status logic into concrete presenters
